CSM, SSS, new widget, Scene inheritance from entt
and a ton of fixes

+++ src/assets.cpp
@@ -36,7 +36,7 @@ std::string TextureAsset::create(const std::string& filepath) {
 
     // TODO: gpu mip mapping, cant right now because assets are loaded in parallel but OpenGL can't do multithreading
 
-    int mipmapLevels = 1 + std::floor(std::log2(std::max(width, height)));
+    int mipmapLevels = 1 + (int)std::floor(std::log2(std::max(width, height)));
 
     for (size_t i = 1; i < mipmapLevels; i++) {
         glm::ivec2 prevSize = { width >> (i - 1), height >> (i - 1) };
